👷‍♂️Include only functional releases into compatibility matrix
